+++ include/LIEF/MachO.hpp
@@ -23,6 +23,7 @@
 #include "LIEF/MachO/Binary.hpp"
 #include "LIEF/MachO/BinaryParser.hpp"
 #include "LIEF/MachO/BindingInfo.hpp"
+#include "LIEF/MachO/BindingInfoIterator.hpp"
 #include "LIEF/MachO/BuildToolVersion.hpp"
 #include "LIEF/MachO/BuildVersion.hpp"
 #include "LIEF/MachO/Builder.hpp"